test 73..

+++ R/BF.cortest.R
@@ -27,24 +27,6 @@
 #'
 #' @examples
 #' \donttest{
-#' # Bayesian correlation analysis of the 6 variables in 'memory' object
-#' # we consider a correlation analysis of the first three variable of the memory data.
-#' #fit <- cor_test(BFpack::memory[,1:3])
-#'
-#' # Bayesian correlation of variables in memory object in BFpack while controlling
-#' # for the Cat variable
-#' #fit <- cor_test(BFpack::memory[,c(1:4)],formula = ~ Cat)
-#'
-#' # Example of Bayesian estimation of polyserial correlations
-#' #memory_example <- memory[,c("Im","Rat")]
-#' #memory_example$Rat <- as.ordered(memory_example$Rat)
-#' #fit <- cor_test(memory_example)
-#'
-#' # Bayesian correlation analysis of first three variables in memory data
-#' # for two different groups
-#' #HC <- subset(BFpack::memory[,c(1:3,7)], Group == "HC")[,-4]
-#' #SZ <- subset(BFpack::memory[,c(1:3,7)], Group == "SZ")[,-4]
-#' #fit <- cor_test(HC,SZ)
 #'
 #' }
 #' @rdname cor_test_continuous
@@ -797,7 +779,7 @@ cor_test <- function(..., formula = NULL, iter = 5e3, burnin = 3e3, nugget.scale
                   WgroupsStore=array(as.double(0),dim=c(samsize0,numG,Ntot,P)),
                   meanMatMeanStore = array(as.double(0),dim=c(samsize0,Ntot,P)),
                   SigmaMatDrawStore = array(as.double(0),dim=c(samsize0,P,P)),
-                  CheckStore = array(as.double(0),dim=c(samsize0,numG,Ntot,P,3*P+2+3))
+                  CheckStore = array(as.double(1),dim=c(samsize0,numG,Ntot,P,3*P+2+3))
   )
 
   varnames <- lapply(1:numG,function(g){
